code refactoring - simplified and improved code styling, spring version update from 3.0.2 to 3.0.4
