--- backend/src/main/java/com/github/advertisement_website/model/response/MessageType.java
@@ -1,6 +1,0 @@
-package com.github.advertisement_website.model.response;
-
-public enum MessageType {
-    SUCCESS,
-    FAILURE
-}
