+++ backend/src/main/java/com/github/advertisement_website/response/AdCategory.java
@@ -1,4 +1,4 @@
-package com.github.advertisement_website.response.adboard;
+package com.github.advertisement_website.response;
 
 public enum AdCategory {
     ELECTRONICS,
