+++ backend/src/main/java/com/github/advertisement_website/response/MessageType.java
@@ -1,0 +1,6 @@
+package com.github.advertisement_website.response;
+
+public enum MessageType {
+    SUCCESS,
+    FAILURE
+}
